feat(api/user): move insert_user_into_db to database module
