+++ api/src/database/user/user.rs
@@ -2,7 +2,7 @@
 
 use anyhow::Context;
 use avina_wire::user::{
-    User, UserClass, UserDetailed, UserMinimal, UserModifyData,
+    User, UserClass, UserCreateData, UserDetailed, UserMinimal, UserModifyData,
 };
 use sqlx::{Executor, FromRow, MySql, Transaction};
 
@@ -493,4 +493,66 @@ pub async fn delete_user_from_db(
         ));
     }
     Ok(())
+}
+
+pub struct NewUser {
+    pub name: String,
+    pub openstack_id: String,
+    pub project_id: u32,
+    pub role: u32,
+    pub is_staff: bool,
+    pub is_active: bool,
+}
+
+// TODO: validate that role is in range 0-2 or 3
+impl TryFrom<UserCreateData> for NewUser {
+    type Error = String;
+
+    // TODO: we might need a more complex function with access to the database
+    //       and the transaction
+    fn try_from(data: UserCreateData) -> Result<Self, Self::Error> {
+        // TODO: really validate data, role range, uuid, string length
+        Ok(Self {
+            name: data.name,
+            openstack_id: data.openstack_id,
+            project_id: data.project,
+            role: data.role.unwrap_or(1),
+            is_staff: data.is_staff.unwrap_or(false),
+            is_active: data.is_active.unwrap_or(true),
+        })
+    }
+}
+
+#[tracing::instrument(
+    name = "insert_user_into_db",
+    skip(new_user, transaction)
+)]
+pub async fn insert_user_into_db(
+    transaction: &mut Transaction<'_, MySql>,
+    new_user: &NewUser,
+) -> Result<u64, MinimalApiError> {
+    // TODO: MariaDB 10.5 introduced INSERT ... RETURNING
+    let query = sqlx::query!(
+        r#"
+        INSERT IGNORE INTO user_user (name, openstack_id, project_id, role, is_staff, is_active)
+        VALUES (?, ?, ?, ?, ?, ?)
+        "#,
+        new_user.name,
+        new_user.openstack_id,
+        new_user.project_id,
+        new_user.role,
+        new_user.is_staff,
+        new_user.is_active,
+    );
+    let result = transaction
+        .execute(query)
+        .await
+        .context("Failed to execute insert query")?;
+    if result.rows_affected() == 0 {
+        return Err(MinimalApiError::ValidationError(
+            "Failed to insert new user, a conflicting entry exists".to_string(),
+        ));
+    }
+    let id = result.last_insert_id();
+    Ok(id)
 }
