+++ api/src/routes/user/import.rs
@@ -20,16 +20,13 @@ use crate::{
         },
         user::{
             project::select_all_projects_from_db,
-            user::select_all_users_from_db,
+            user::{NewUser, insert_user_into_db, select_all_users_from_db},
         },
     },
     error::NormalApiError,
     ldap::AvinaLdap,
     openstack::OpenStack,
-    routes::{
-        project::create::{NewProject, insert_project_into_db},
-        user::user::create::{NewUser, insert_user_into_db},
-    },
+    routes::project::create::{NewProject, insert_project_into_db},
     startup::AvinaLdapConfig,
 };
